YOLO/requirements.txt
alembic==1.17.2
attrs==25.4.0
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
click-plugins==1.1.1.2
cligj==0.7.2
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.0
fiona==1.10.1
Flask==3.1.2
flask-cors==6.0.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.60.1
fsspec==2025.9.0
geopandas==1.1.1
greenlet==3.2.4
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
pillow==12.0.0
polars==1.34.0
polars-runtime-32==1.34.0
psutil==7.1.1
psycopg2-binary==2.9.11
pyogrio==0.11.1
pyparsing==3.2.5
pyproj==3.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.2
six==1.17.0
SQLAlchemy==2.0.44
sympy==1.14.0
torch==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.221
ultralytics-thop==2.0.17
urllib3==2.5.0
Werkzeug==3.1.3
